load fix files

+++ src/utils.js
@@ -4,8 +4,12 @@ export const getFromStorage = function (key) {
 
 export const addToStorage = function (obj, key) {
   const storageData = getFromStorage(key);
-  storageData.push(obj);
-  localStorage.setItem(key, JSON.stringify(storageData));
+  // Проверяем, существует ли уже пользователь с таким логином 
+  const existingUserIndex = storageData.findIndex(user => user.login === obj.login);
+  if (existingUserIndex === -1) {
+    storageData.push(obj);
+    localStorage.setItem(key, JSON.stringify(storageData));
+  }
 };
 
 export const generateTestUser = function (User) {
